Combined certs

+++ images/admin_api/src/components/digitalTwin/digitalTwinDAL.ts
@@ -1027,8 +1027,9 @@ export const deleteBucketFiles = async (filesToRemove: string[]) => {
 
 export const removeFilesFromBucketFolder = async (folderPath: string) => {
 	const filesToRemove = await getBucketFolderFileList(folderPath);
-	console.log("filesToRemove=", filesToRemove)
-	await deleteBucketFiles(filesToRemove);
+	if (filesToRemove.length !== 0) {
+		await deleteBucketFiles(filesToRemove);
+	}
 }
